% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaFunction.R
\name{formulaHandler_I}
\alias{formulaHandler_I}
\title{Function to handle I() in LINPRED}
\usage{
formulaHandler_I(x, defaultBracket, coefPrefix, sdPrefix, modelInfo, env, ...)
}
\arguments{
\item{x}{A formulaComponentFunction object created from an I() term}

\item{defaultBracket}{The bracket from the LHS of LINPRED}

\item{coefPrefix}{The prefix to use for any new linear predictor parameters created}

\item{sdPrefix}{The prefix to use for any new standard deviation parameters created}

\item{modelInfo}{Named list containing model information including constants}

\item{env}{Environment in which the LINPRED macro was called}

\item{...}{Not currently used}
}
\value{
An object of class \code{formulaComponentFixed}.
}
\description{
Translates I() in an R formula passed to LINPRED into corresponding
NIMBLE code (and new constant) for a scaled covariate.
Only allows for expressions involving one covariate (not functions of covariates).
This is used internally by \code{LINPRED} and should not be called directly. 
New formula functions should have the same arguments, naming structure, class 
(\code{nimbleFormulaHandler}) and return object class (\code{formulaComponent}).
}
