% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{formulaHandler_ICAR}
\alias{formulaHandler_ICAR}
\title{Function to handle ICAR() in LINPRED}
\usage{
formulaHandler_ICAR(
  x,
  defaultBracket,
  coefPrefix,
  sdPrefix,
  modelInfo,
  env,
  ...
)
}
\arguments{
\item{x}{A \code{formulaComponentFunction} object created from an ICAR() term}

\item{defaultBracket}{The bracket from the LHS of LINPRED}

\item{coefPrefix}{The prefix to use for any new linear predictor parameters created}

\item{sdPrefix}{The prefix to use for any new standard deviation parameters created}

\item{modelInfo}{Named list containing model information including constants}

\item{env}{Environment in which the LINPRED macro was called}

\item{...}{Not currently used}
}
\value{
An object of class \code{formulaComponent}.
}
\description{
Translates ICAR() in an R formula passed to LINPRED into corresponding
NIMBLE code for an ICAR spatial random effect. This is used internally
by \code{LINPRED} and should not be called directly. New formula functions
should have the same arguments, naming structure, class (\code{nimbleFormulaHandler}) 
and return object class (\code{formulaComponent}).
}
\author{
Ken Kellner
}
