% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialMacros.R
\name{GaussianProcess}
\alias{GaussianProcess}
\title{Macro to build code for an Gaussian Process random effect}
\arguments{
\item{spatData}{Spatial data object; can be SpatialPoints, SpatialPolygons,
or data frame of coordinates}

\item{prefix}{All macro-generated parameters will begin with this prefix,
default is GP_ (so x becomes GP_x)}
}
\description{
Takes spatial data as a Spatial object or data frame of coordinates, and
generates corresponding code for an Gaussian process random effect
}
\author{
Ken Kellner
}
