% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{GP}
\alias{GP}
\title{Macro to build code for an Gaussian Process random effect}
\arguments{
\item{spatData}{Spatial data object; can be SpatialPoints, SpatialPolygons,
or data frame of coordinates}

\item{prefix}{All macro-generated parameters will begin with this prefix,
default is GP_ (so x becomes GP_x)}

\item{priorSpecs}{List of prior specifications, should be generated using 
setPriors()}
}
\description{
Takes spatial data as a Spatial object or data frame of coordinates, and
generates corresponding code for an Gaussian process random effect
}
\author{
Ken Kellner
}
