% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{ICAR}
\alias{ICAR}
\title{Macro to build code for an ICAR random effect}
\arguments{
\item{spatData}{Spatial data object; can be SpatialPoints, SpatialPolygons,
or data frame of coordinates}

\item{threshold}{Distance at and below which two points will be considered
neighbors}

\item{prefix}{All macro-generated parameters will begin with this prefix,
default is ICAR_ (so x becomes ICAR_x)}

\item{zero_mean}{Value passed to corresponding argument in dcar_normal() 
distribution in NIMBLE code}
}
\description{
Takes spatial data as a Spatial object or data frame of coordinates, and
generates corresponding code for an ICAR random effect
}
\author{
Ken Kellner
}
