% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaFunction.R
\name{IFormulaFunction}
\alias{IFormulaFunction}
\title{Function to handle I() in LINPRED}
\arguments{
\item{x}{A formulaComponentFunction object created from an I() term}
}
\description{
Translates I() in an R formula passed to LINPRED into corresponding
NIMBLE code (and new constant) for a scaled covariate.
Only allows for expressions involving one covariate (not functions of covariates).
}
