% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linPred.R
\name{priors}
\alias{priors}
\title{Macro to build code for priors on a linear predictor from R formula}
\arguments{
\item{formula}{An R formula The formula must be right-hand side only (e.g. ~x). 
This must always be the first argument supplied to priors.}

\item{coefPrefix}{All model coefficient names will begin with this prefix.
default is beta_ (so x becomes beta_x, etc.)}

\item{sdPrefix}{All dispersion parameters will begin with this prefix.
default is no prefix.}

\item{priorSettings}{List of prior specifications, should be generated using 
setPriors()}

\item{modMatNames}{Logical, should parameters be named so they match the
names you would get from R's model.matrix function?}
}
\description{
Generates appropriate priors for a linear predictor derived from an 
R formula. As such it makes the most sense to use this macro together with
the linPred macro which takes similar arguments.
}
\examples{
\donttest{
constants <- list(x = rnorm(10), 
                  x2 = factor(sample(letters[1:3], 10, replace=TRUE)))

# Just linear predictor
code <- nimbleCode({
  y[1:n] ~ linPred(~x + x2)
  priors(~x + x2, coefPrior = dnorm(0, sd = 5), modMatNames=TRUE)
 })
nimble:::processModelMacros(code, constants)$code
}
}
\author{
Ken Kellner
}
