% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaFunction.R
\name{scaleFormulaFunction}
\alias{scaleFormulaFunction}
\title{Function to handle scale() in LINPRED}
\arguments{
\item{x}{A formulaComponentFunction object created from a scale() term}
}
\description{
Translates scale() in an R formula passed to LINPRED into corresponding
NIMBLE code (and new constant) for a scaled covariate.
}
