% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simplifyForLoops}
\alias{simplifyForLoops}
\title{Simplify for loop structure in NIMBLE model code}
\usage{
simplifyForLoops(code, new_indices = NULL)
}
\arguments{
\item{code}{NIMBLE code for a model, such as from the output of model$getCode()}

\item{new_indices}{A list of new for loop indices that will replace the existing
indices. The new indices must be quoted values (i.e., "names"/symbols).
If NULL, letters starting with 'i' will be used. If FALSE, no indices will
be replaced.}
}
\value{
Model code with collapsed/simplified for loops and indices.
}
\description{
Takes the code for a NIMBLE model and attempts to combine for loops
that share the same index range, in order to simplify the code
structure. Optionally, can also replace existing for loop indices
with a (potentially smaller, simpler) set of new indices.
This function is particularly useful for simplifying code generated by
macros, which often creates many for loops with the same indices and 
uses complex indices like 'i_1', 'i_2', etc. which could be simplified to
'i', 'j', etc.
}
\author{
Ken Kellner
}
