% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LINPRED.R
\name{uppertri_mult_diag}
\alias{uppertri_mult_diag}
\title{uppertri_mult_diag}
\usage{
uppertri_mult_diag(mat, vec)
}
\arguments{
\item{mat}{upper triangular Cholesky factor of correlation matrix (Ustar)}

\item{vec}{vector of standard deviations for individual random effects}
}
\description{
nimbleFunction needed when fitting correlated random effects.
Generates upper triangular Cholesky factor of covariance matrix (U in code)
from upper tri Cholesky factor of correlation matrix (Ustar in code)
and vector of standard deviations. Taken from the NIMBLE manual, 
section 5.2.4.1.2 LKJ distribution for correlation matrices.
}
