% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forLoop.R
\name{forLoop}
\alias{forLoop}
\title{Macro to build for loop(s) from code with index ranges in brackets}
\arguments{
\item{code}{The right-hand side of a parameter declaration}
}
\description{
This macro takes a line of BUGS code with index ranges inside brackets
on either the left-hand side of a declaration or both the left- and
right-hand sides of a declaration and constructs a corresponding
for loop or series of nested for loops.
}
\examples{
code <- nimbleCode({
   y[1:n, 1:2, 1] ~ forLoop(dnorm(mu[1:n], sigma))
   mu[1:n] <- forLoop(beta[1] + beta[2]*x[1:n])
})

mod <- nimbleModel(code, constants=list(n=10))
mod$getCode()
}
\author{
Ken Kellner
}
