% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LINPRED.R
\name{LINPRED_PRIORS}
\alias{LINPRED_PRIORS}
\title{Macro to build code for priors on a linear predictor from an R formula}
\arguments{
\item{formula}{An R formula The formula must be right-hand side only (e.g., \code{~x}). 
This must always be the first argument supplied to \code{LINPRED_PRIORS}.}

\item{coefPrefix}{All model coefficient names will begin with this prefix.
default is \code{"beta_"} (so 'x' becomes 'beta_x', etc.)}

\item{sdPrefix}{All dispersion parameters will begin with this prefix.
Default is no prefix.}

\item{priorSpecs}{List of prior specifications, generated using \code{setPriors}.
setPriors()}

\item{modMatNames}{Logical indicating if parameters should be named so they match the
names one would get from R's \code{model.matrix}. Default is \code{FALSE}.}

\item{noncenter}{Logical indicating whether to use noncentered parameterization.
Default is \code{FALSE}.}

\item{centerVar}{Grouping variable (covariate) to 'center' the random effects on. By
default all random effects have mean 0 as with \code{lme4}.}

\item{modelInfo}{Used internally by nimbleMacros; a list of model information such as constants and dimensions}

\item{.env}{Used internally by nimbleMacros; the environment where the model was created}
}
\description{
Generates appropriate priors for a linear predictor derived from an 
R formula. As such it makes the most sense to use this macro together with
the LINPRED macro, which takes similar arguments.
}
\examples{
constants <- list(x = rnorm(3), x2 = factor(letters[1:3]))
code <- nimbleCode({
  LINPRED_PRIORS(~x + x2)
})

mod <- nimbleModel(code, constants = constants)
mod$getCode()
}
\author{
Ken Kellner
}
