% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LINPRED.R
\name{LINPRED_PRIORS}
\alias{LINPRED_PRIORS}
\title{Macro to build code for priors on a linear predictor from R formula}
\arguments{
\item{formula}{An R formula The formula must be right-hand side only (e.g. ~x). 
This must always be the first argument supplied to LINPRED_PRIORS}

\item{coefPrefix}{All model coefficient names will begin with this prefix.
default is beta_ (so x becomes beta_x, etc.)}

\item{sdPrefix}{All dispersion parameters will begin with this prefix.
default is no prefix.}

\item{priorSpecs}{List of prior specifications, should be generated using 
setPriors()}

\item{modMatNames}{Logical, should parameters be named so they match the
names you would get from R's model.matrix function?}

\item{noncenter}{Logical, use noncentered parameterization?}

\item{centerVar}{Grouping covariate to 'center' on in parameterization. By
default all random effects have mean 0 as with lme4.}
}
\description{
Generates appropriate priors for a linear predictor derived from an 
R formula. As such it makes the most sense to use this macro together with
the LINPRED macro which takes similar arguments.
}
\examples{
constants <- list(x = rnorm(3), x2 = factor(letters[1:3]))
code <- nimbleCode({
  LINPRED_PRIORS(~x + x2)
})

mod <- nimbleModel(code, constants = constants)
mod$getCode()
}
\author{
Ken Kellner
}
