% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaFunction.R
\name{offsetFormulaFunction}
\alias{offsetFormulaFunction}
\title{Function to handle offset() in LINPRED}
\arguments{
\item{x}{A formulaComponentFunction object created from an offset() term}
}
\description{
Translates offset() in an R formula passed to LINPRED into corresponding
NIMBLE code for an offset in the linear predictor.
}
