% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors}
\alias{priors}
\title{Macro to build code for priors on a linear predictor from R formula}
\arguments{
\item{formula}{An R formula The formula must be right-hand side only (e.g. ~x). 
This must always be the first argument supplied to priors.}

\item{sdPrefix}{All dispersion parameters will begin with this prefix.
default is no prefix.}

\item{coefPrior}{BUGS code for prior on coefficients. Default is dnorm(0, sd=10).}

\item{sdPrior}{BUGS code for prior on dispersion parameters. Default is
half-Cauchy T(dt(0, 0.1, 1), 0,).}

\item{modMatNames}{Logical, should parameters be named so they match the
names you would get from R's model.matrix function?}
}
\description{
Generates appropriate priors for a linear predictor derived from an 
R formula. As such it makes the most sense to use this macro together with
the linPred macro which takes similar arguments.
}
\examples{
\donttest{
constants <- list(x = rnorm(10), 
                  x2 = factor(sample(letters[1:3], 10, replace=T)))

# Just linear predictor
code <- nimbleCode({
  y[1:n] ~ linPred(~x + x2)
  beta. ~ priors(~x + x2, coefPrior = dnorm(0, sd = 5), modMatNames=TRUE)
 })
nimble:::codeProcessModelMacros(code, constants)$code
}
}
\author{
Ken Kellner
}
