% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaFunction.R
\name{scaleFormulaFunction}
\alias{scaleFormulaFunction}
\title{Function to handle scale() in LINPRED}
\usage{
scaleFormulaFunction(
  x,
  defaultBracket,
  coefPrefix,
  sdPrefix,
  modelInfo,
  env,
  ...
)
}
\arguments{
\item{x}{A formulaComponentFunction object created from a scale() term}

\item{defaultBracket}{The bracket from the LHS of LINPRED}

\item{coefPrefix}{The prefix to use for any new linear predictor parameters created}

\item{sdPrefix}{The prefix to use for any new standard deviation parameters created}

\item{modelInfo}{Named list containing model information including constants}

\item{env}{Environment in which the LINPRED macro was called}

\item{...}{Not currently used}
}
\value{
An object of class \code{formulaComponentFixed}.
}
\description{
Translates scale() in an R formula passed to LINPRED into corresponding
NIMBLE code (and new constant) for a scaled covariate. This is used internally
by \code{LINPRED} and should not be called directly. New formula functions
should have the same arguments, naming structure, class (\code{formulaFunction}) 
and return object class (\code{formulaComponent}).
}
